python-examples/hvac/requirements.txt
google-api-python-client
google-auth
hvac
